+++ cmd/setup-node/commands/root_windows.go
@@ -1,0 +1,11 @@
+//+build windows
+
+package commands
+
+import (
+	"github.com/SkycoinProject/skycoin/src/util/logging"
+)
+
+func setupSyslog(_ *logging.Logger) {
+
+}
