+++ cmd/skywire-visor/commands/root.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"io/ioutil"
 	"log"
-	"log/syslog"
 	"net/http"
 	_ "net/http/pprof" // nolint:gosec // TODO: consider removing for security reasons
 	"os"
@@ -24,7 +23,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/SkycoinProject/skywire-mainnet/internal/utclient"
-	skywirelogging "github.com/SkycoinProject/skywire-mainnet/pkg/logging"
 	"github.com/SkycoinProject/skywire-mainnet/pkg/restart"
 	"github.com/SkycoinProject/skywire-mainnet/pkg/util/buildinfo"
 	"github.com/SkycoinProject/skywire-mainnet/pkg/util/pathutil"
@@ -123,23 +121,6 @@ func (cfg *runCfg) startProfiler() *runCfg {
 	}
 
 	cfg.profileStop = profile.Start(profile.ProfilePath("./logs/"+cfg.tag), option).Stop
-
-	return cfg
-}
-
-func (cfg *runCfg) startLogger() *runCfg {
-	cfg.masterLogger = logging.NewMasterLogger()
-	cfg.logger = cfg.masterLogger.PackageLogger(cfg.tag)
-
-	if cfg.syslogAddr != "none" {
-		hook, err := skywirelogging.NewSysLogHook("udp", cfg.syslogAddr, syslog.LOG_INFO, cfg.tag)
-		if err != nil {
-			cfg.logger.Error("Unable to connect to syslog daemon:", err)
-		} else {
-			cfg.masterLogger.AddHook(hook)
-			cfg.masterLogger.Out = ioutil.Discard
-		}
-	}
 
 	return cfg
 }
